feat(ltk-manager): clean up library backend
