+++ crates/ltk-manager/src/modules/library/api/useModThumbnail.ts
@@ -1,5 +1,4 @@
-import { skipToken, useQuery } from "@tanstack/react-query";
-import { convertFileSrc } from "@tauri-apps/api/core";
+import { useQuery } from "@tanstack/react-query";
 
 import { api, type AppError } from "@/lib/tauri";
 import { queryFn } from "@/utils/query";
@@ -7,13 +6,13 @@ import { queryFn } from "@/utils/query";
 import { libraryKeys } from "./keys";
 
 /**
- * Hook to fetch and cache mod thumbnail absolute path, returning a Tauri asset URL.
+ * Hook to fetch a mod thumbnail as a base64 data URL, loaded on-the-fly from the archive.
+ * Returns `null` if the mod has no thumbnail.
  */
-export function useModThumbnail(modId: string, thumbnailPath?: string) {
-  return useQuery<string, AppError, string>({
-    queryKey: libraryKeys.thumbnail(modId, thumbnailPath),
-    queryFn: thumbnailPath ? queryFn(() => api.getModThumbnail(thumbnailPath)) : skipToken,
-    select: (path) => (path ? convertFileSrc(path) : ""),
-    staleTime: Infinity, // Thumbnails don't change often
+export function useModThumbnail(modId: string) {
+  return useQuery<string | null, AppError>({
+    queryKey: libraryKeys.thumbnail(modId),
+    queryFn: queryFn(() => api.getModThumbnail(modId)),
+    staleTime: Infinity,
   });
 }
